Add upper bound for parse-duration package version (#97)
NPM package
[parse-duration](https://www.npmjs.com/package/parse-duration) version
was bumped from `1.1.0` to `2.1.3` In PR #95 for CodePush CLI. This
caused the build to fail with the following error:

```shell
% npm run build

> code-push-cli@0.0.1 build
> tsc

script/command-parser.ts:1321:21 - error TS2349: This expression is not callable.
  Type 'typeof import("/private/tmp/ms-code-push-server/cli/node_modules/parse-duration/index")' has no call signatures.

1321   return Math.floor(parseDuration(durationString));
                         ~~~~~~~~~~~~~


Found 1 error in script/command-parser.ts:1321
```

As current code-base is not compatible with `parse-duration` versions
`2.0.0+`, restrict `parse-duration` to use version `1.x.y`.